Fixed methods called from wrong thread, other code cleanup / improvements

+++ app/src/main/java/org/linphone/ui/call/view/VuMeterView.kt
@@ -82,7 +82,7 @@ class VuMeterView : View {
     }
 
     private fun createShader(): Shader {
-        val level = (height - height * vuMeterPercentage).toFloat()
+        val level = (height - height * vuMeterPercentage)
 
         val bitmap = createBitmap(width, height)
         val canvas = Canvas(bitmap)
